+++ facia-tool/cloudformation/frontend.json
@@ -57,15 +57,6 @@
             {
                 "Effect": "Allow",
                 "Action": [
-                  "cloudwatch:GetMetricStatistics",
-                  "cloudwatch:ListMetrics",
-                  "cloudwatch:PutMetricData"
-                ],
-                "Resource": "*"
-            },
-            {
-                "Effect": "Allow",
-                "Action": [
                   "sqs:SendMessage",
                   "sqs:ReceiveMessage",
                   "sqs:DeleteMessage"
